Calculate tolerable discharge

Uses the site data to calculate a natural stormwater run-off for a yearly rain event

calculate_tolerable_discharge(
  area_catch = 1,
  area_con_catch = 1,
  area_plan = 0,
  slope_catch = 0.1,
  Hq1pnat_catch = NULL,
  Hq2pnat_catch = NULL,
  site_data = NULL,
  verbose = TRUE
)

Arguments

area_catch
catchment area in km²

area_con_catch
impervious catchment area discharging into the surface water in km²

area_plan
planning area in km² (default is 0 -> no planning area)

slope_catch
average slope of the catchment area in % (Defalut is 0.1)

Hq1pnat_catch
natural average catchment discharge for a yearly rain event in L/(s*km²) (Defautl is NULL)

Hq2pnat_catch
natural average catchment discharge for a bienneal rain event in L/(s*km²) (Defautl is NULL)

site_data
site data loaded by function "load_site_data". If defined, this overwrites all other values.

verbose
if TRUE returns results as informative messages, If FALSE only return numeric value for planning area.

Value

Table with tolerable discharge for the whole Catchment and planning area in L/s. Furthermore, x is given which is a factor for allowed discharge increase compared to the natural status and is included in the calculation for the tolerable discharges. The definition of x can be found in guideline DWA-A 102-3

Examples