get_rain

The rate of the yearly rain event depends on the prescribed duration. In this function the duration is either calculated using the natural catchment discharge, using the average river flow or entered manualle.

get_rain(
  area_catch,
  river_cross_section,
  river_length,
  x_coordinate,
  y_coordinate,
  Hq_pnat1_catch = NULL,
  slope = 0.1,
  use_p1nat = TRUE,
  river_flow = NULL,
  mins = NULL
)

Arguments

area_catch

The catchment area in km2

river_cross_section

The average river cross section in the catchment in m2

river_length

The length of the affected urban river stretch in m

x_coordinate, y_coordinate

coordinates in ETRS89. See Datails for more information.

Hq_pnat1_catch

the natural catchment discharge for a yearly rain event in L/(s*km2). If NULL it will be estimated by slope and area of the catchment

slope

Average slope of the catchment in % (Default is 0.1)

use_p1nat

If TRUE, the natural catchment discharge is used (see get_Hq1_pnat) is used to define the precipitation duration. If FALSE the average river flow is used. Exception: If mins is defined, this value is used.

river_flow

The average river flow in m³/s (only needed if use_p1nat = FALSE and min = NULL)

mins

The Default is NULL. In this case either natural catchment discharge or average river flow is used for precipitation duration. If not Null, mins is used and overwrites the parameter "use_p1nat".

Value

A Vector with the calculated duration of precipitation in minutes and the rain event based on KOSTRA in L/(s*ha)

Details

The KOSTRA Data is available for a grid of x x x km. The location is given in coordinates in the ETRS89 system (For information see: https://epsg.io/3034) Longitudes and Lattidudes in WGS84 can be converted into ETRS89 here: https://epsg.io/transform#s_srs=4326&t_srs=3034