Loading local background concentration

This functions loads the data from the sheet "pollution_data" within the R2Q-Excel file for data entry

load_background_data(
  data.dir,
  filename,
  default_for_na = TRUE,
  SUW_type = "river"
)

Arguments

data.dir

The directory of the entry data table.

filename

Name of the R2Q-Excel File including ".xlsx".

default_for_na

If TRUE, default values are used for substances that were not measured

SUW_type

Only used if default_for_na is TRUE. "lake" or "river", "river" is used as default

Value

A data frame background concentration as defined in the Excel sheet. If default values are used this is documented in the "comment" column.

Examples

load_site_data(
data.dir = system.file("extdata/Data_entry", package = "r2q"), 
filename = "Baukau.xlsx")
#> $City_name
#> $City_name$Unit
#> [1] NA
#> 
#> $City_name$Value
#> [1] "Herne"
#> 
#> $City_name$Explanation
#> [1] "Name of the location"
#> 
#> $City_name$Obligatory
#> [1] NA
#> 
#> 
#> $x_coordinate
#> $x_coordinate$Unit
#> [1] NA
#> 
#> $x_coordinate$Value
#> [1] 3813634
#> 
#> $x_coordinate$Explanation
#> [1] "x value in ETRS 89 - Coordinate System"
#> 
#> $x_coordinate$Obligatory
#> [1] "x"
#> 
#> 
#> $y_coordinate
#> $y_coordinate$Unit
#> [1] NA
#> 
#> $y_coordinate$Value
#> [1] 2753912
#> 
#> $y_coordinate$Explanation
#> [1] "y value in ETRS 89 - Coordinate System"
#> 
#> $y_coordinate$Obligatory
#> [1] "x"
#> 
#> 
#> $SUW_name
#> $SUW_name$Unit
#> [1] NA
#> 
#> $SUW_name$Value
#> [1] "Ostbach"
#> 
#> $SUW_name$Explanation
#> [1] "Name of surface water body"
#> 
#> $SUW_name$Obligatory
#> [1] NA
#> 
#> 
#> $SUW_type
#> $SUW_type$Unit
#> [1] NA
#> 
#> $SUW_type$Value
#> [1] "river"
#> 
#> $SUW_type$Explanation
#> [1] "river or lake?"
#> 
#> $SUW_type$Obligatory
#> [1] NA
#> 
#> 
#> $LAWA_type
#> $LAWA_type$Unit
#> [1] NA
#> 
#> $LAWA_type$Value
#> [1] 19
#> 
#> $LAWA_type$Explanation
#> [1] "LAWA type of surface water body"
#> 
#> $LAWA_type$Obligatory
#> [1] NA
#> 
#> 
#> $Q_mean
#> $Q_mean$Unit
#> [1] "m3/s"
#> 
#> $Q_mean$Value
#> [1] 0.04
#> 
#> $Q_mean$Explanation
#> [1] "Average annual river runoff MQ"
#> 
#> $Q_mean$Obligatory
#> [1] "x"
#> 
#> 
#> $river_cross_section
#> $river_cross_section$Unit
#> [1] "m2"
#> 
#> $river_cross_section$Value
#> [1] 0.54
#> 
#> $river_cross_section$Explanation
#> [1] "average river cross section"
#> 
#> $river_cross_section$Obligatory
#> [1] "x"
#> 
#> 
#> $river_length
#> $river_length$Unit
#> [1] "m"
#> 
#> $river_length$Value
#> [1] 5000
#> 
#> $river_length$Explanation
#> [1] "length of affected urban river stretch"
#> 
#> $river_length$Obligatory
#> [1] "x"
#> 
#> 
#> $impact_time
#> $impact_time$Unit
#> [1] "min"
#> 
#> $impact_time$Value
#> [1] 1080
#> 
#> $impact_time$Explanation
#> [1] "flow_time rounded to next rain duration class"
#> 
#> $impact_time$Obligatory
#> [1] NA
#> 
#> 
#> $rain_year
#> $rain_year$Unit
#> [1] "mm/yr"
#> 
#> $rain_year$Value
#> [1] 822
#> 
#> $rain_year$Explanation
#> [1] "annual rainfall [mm/yr]"
#> 
#> $rain_year$Obligatory
#> [1] NA
#> 
#> 
#> $area_catch
#> $area_catch$Unit
#> [1] "km2"
#> 
#> $area_catch$Value
#> [1] 5.262
#> 
#> $area_catch$Explanation
#> [1] "river catchment area"
#> 
#> $area_catch$Obligatory
#> [1] "x"
#> 
#> 
#> $area_con_catch
#> $area_con_catch$Unit
#> [1] "km2"
#> 
#> $area_con_catch$Value
#> [1] 1.651
#> 
#> $area_con_catch$Explanation
#> [1] "connected impervious area in river catchment"
#> 
#> $area_con_catch$Obligatory
#> [1] "x"
#> 
#> 
#> $f_D_catch
#> $f_D_catch$Unit
#> [1] "-"
#> 
#> $f_D_catch$Value
#> [1] 0.88
#> 
#> $f_D_catch$Explanation
#> [1] "Average runoff coefficient of impervious area of river catchment"
#> 
#> $f_D_catch$Obligatory
#> [1] "x"
#> 
#> 
#> $Hq1pnat_catch
#> $Hq1pnat_catch$Unit
#> [1] "l/(s*kmì)"
#> 
#> $Hq1pnat_catch$Value
#> [1] 130
#> 
#> $Hq1pnat_catch$Explanation
#> [1] "Natural run-off of urban area (\"area_catch\") for a yearly rain event"
#> 
#> $Hq1pnat_catch$Obligatory
#> [1] NA
#> 
#> 
#> $separate_catch
#> $separate_catch$Unit
#> [1] "%"
#> 
#> $separate_catch$Value
#> [1] 25
#> 
#> $separate_catch$Explanation
#> [1] "share of impervious area connected to SUW via separate sewer system"
#> 
#> $separate_catch$Obligatory
#> [1] NA
#> 
#> 
#> $slope_catch
#> $slope_catch$Unit
#> [1] "%"
#> 
#> $slope_catch$Value
#> [1] 0.1
#> 
#> $slope_catch$Explanation
#> [1] "Average slope of the catchment area"
#> 
#> $slope_catch$Obligatory
#> [1] NA
#> 
#> 
#> $area_plan
#> $area_plan$Unit
#> [1] "km2"
#> 
#> $area_plan$Value
#> [1] 1.556295
#> 
#> $area_plan$Explanation
#> [1] "planing area"
#> 
#> $area_plan$Obligatory
#> [1] "x"
#> 
#> 
#> $area_con_plan
#> $area_con_plan$Unit
#> [1] "km2"
#> 
#> $area_con_plan$Value
#> [1] 0.548919
#> 
#> $area_con_plan$Explanation
#> [1] "connected impervious area in planing area"
#> 
#> $area_con_plan$Obligatory
#> [1] "x"
#> 
#> 
#> $f_D_plan
#> $f_D_plan$Unit
#> [1] NA
#> 
#> $f_D_plan$Value
#> [1] 0.88
#> 
#> $f_D_plan$Explanation
#> [1] "Average runoff coefficient of impervious area of planing area"
#> 
#> $f_D_plan$Obligatory
#> [1] "x"
#> 
#> 
#> $areaType
#> residential_suburban     residential_city             industry 
#>         4.000000e-01         4.000000e-01         2.000000e-01 
#>               street            connected 
#>        -2.220446e-16         3.000000e-01 
#>