calculate maximal allowable connected area in a river catchment at steady state

based on concentration thresholds based on DWA A102 approach can be used for annual and acute impacts if steady state concentration can be assumed or if inlets are at one inlet point on the river stretch

max_area_steady_state(
  Q_river,
  Ci_river,
  Ci_threshold,
  Ci_storm,
  coeff_runoff,
  Q_rain,
  substance_i = NULL
)

Arguments

Q_river
Annual river flow in m³/a

Ci_river
Background concentration for substance i. Concentration unit must fit to Ci_threshold and Ci_storm.

Ci_threshold
Threshold value for substance i. Concentration unit must fit to Ci_river and Ci_storm.

Ci_storm
Concentration in stormwater run-off for substance i. Concentration unit must fit to Ci_threshold and Ci_river.

coeff_runoff
Run-off coefficient of connected impervious area

Q_rain
Annual amount of rain amount in mm/a

substance_i
the substance name is only used for messages, not for calculation

Value

maximal connected impervious area in km2