Loading local background concentration

This functions loads the data from the sheet "pollution_data" within the R2Q-Excel file for data entry

load_background_data(
  data.dir,
  filename,
  default_for_na = TRUE,
  SUW_type = "river"
)

Arguments

data.dir
The directory of the entry data table.

filename
Name of the R2Q-Excel File including ".xlsx".

default_for_na
If TRUE, default values are used for substances that were not measured

SUW_type
Only used if default_for_na is TRUE. "lake" or "river", "river" is used as default

Value

A data frame background concentration as defined in the Excel sheet. If default values are used this is documented in the "comment" column.

Examples

load_site_data(
data.dir = "inst/extdata/Data_entry", 
filename = "Bsp_Herne.xlsx")

#> Error: `path` does not exist: ‘inst/extdata/Data_entry/Bsp_Herne.xlsx’